requirements.txt
chainlit
hugchat
python-dotenv
sentence-transformers 
scipy